remove RPARSERCONTEXTFORPOSITION, fix some int width issues and force re-size for document windows
git-svn-id: https://svn.r-project.org/R-packages/trunk/Mac-GUI@7927 694ef91d-65df-0310-b7bb-92e67a308ead

+++ RTextView.h
@@ -79,7 +79,7 @@ extern BOOL RTextView_autoCloseBrackets;
 
 - (void)setConsoleMode:(BOOL)isConsole;
 
-- (int)  parserContextForPosition:(int)position;
+- (int)  parserContextForPosition:(NSInteger)position;
 - (void) showHelpForCurrentFunction;
 - (void) currentFunctionHint;
 - (BOOL) wrapSelectionWithPrefix:(NSString *)prefix suffix:(NSString *)suffix;
